feat: add filter by price #4
